Introduce m4bber, an auto-m4b rewrite
The `m4bber` tool's lineage is from auto-m4b, but has different goals.
Rather than be a standalone daemon that runs on a cycle, `m4bber`'s
goal is to simply process the files in the requested directory when
called, then exits when all the files are processed. This allows
greater ability to employ this functionality in different contexts
(e.g., as part of a cron job, or triggered by a filesystem event).

Likewise, the Dockerfile is streamlined to use a multistage build,
which helps keep the final image's size to a minimum. It also uses
upgrades the base image to Ubuntu 24, and relies on standard packages
rather than custom builds, when possible.

The script was rewritten to rely on command-line flags rather than
(solely) on environment variables to customize behavior.